% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expertise_diversity.R
\name{check_expertise_diversity}
\alias{check_expertise_diversity}
\title{Check Expertise Diversity Across Groups}
\usage{
check_expertise_diversity(students)
}
\arguments{
\item{students}{A data frame.}
}
\value{
Return a message and the value TRUE if gender distribution is OK, FALSE otherwise.
}
\description{
After groups have been made, check expertise distribution.
}
\examples{
# Check expertise diversity on the example dataset included in the makegroups package
check_expertise_diversity(make_groups(students, 3))
}
