% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gender_balance.R
\name{check_gender_balance}
\alias{check_gender_balance}
\title{Check Gender Balance in a Group}
\usage{
check_gender_balance(students)
}
\arguments{
\item{students}{A data frame.}
}
\value{
Return a message and the value TRUE if gender distribution is OK, FALSE otherwise.
}
\description{
After groups have been made, check gender balance.
}
\examples{
# Check gender balance on the example dataset included in the makegroups package
check_gender_balance(make_groups(students, 3))
}
