% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make Groups}
\usage{
make_groups(students, group_size)
}
\arguments{
\item{students}{A data frame.}

\item{group_size}{A number. integer}
}
\value{
The students data frame with a new column \code{group}.
}
\description{
Given a list of students and a desired group size, assign the students to
groups in such a way that the diversity of group members is maximised.
}
\examples{
# Dummy data included in the groupmaker package
students

# How many members should a group have?
group_size <- 3

# Make groups of students
make_groups(students = students, group_size = group_size)
}
