% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Example data: Students}
\format{
A data frame with 7 rows and 4 variables
\describe{
\item{name}{Name of the student}
\item{nationality}{Natinoality of the student}
\item{designer}{A logical variable: TRUE if the student is a designer, FALSE otherwise}
\item{gender}{Gender of the student}
}
}
\usage{
students
}
\description{
A dummy dataset with fictional students used to illustrate
the functionality of the groupmaker package
}
\keyword{datasets}
