% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make Groups}
\usage{
make_groups(students, group_size)
}
\arguments{
\item{students}{data frame}

\item{group_size}{integer}
}
\value{

}
\description{
Given a list of students and a desired group size, assign students to
groups in such a way that diversity of group members is maximised.
}
\examples{
require(dplyr)
students <- tibble(name = c("John", "Mary", "Bob", "Kate", "Ganesh", "Marta", "Janneke"),
                   nationality = c("NL", "RO", "US", "NL", "IN", "DE", "NL"),
                   designer = c(TRUE, FALSE, FALSE, TRUE, TRUE, FALSE, FALSE),
                   gender = c("M", "F", "M", "F", "M", "F", "F"))
students$nationality <- as.factor(students$nationality)
students$gender <- as.factor(students$gender)
make_groups(students, 3)
}
